refactor: merge BroserWallet and ChannelWallet
