--- packages/xstate-wallet/src/browser-wallet.ts
@@ -1,1 +1,0 @@
-export {ChannelWallet2, Message} from './channel-wallet2';
